# [v1.9.0] [zos_copy] Enhancement/764/copy members