##### SUMMARY
<!--- Describe the change below, including rationale and design decisions -->
Improved the performance of copying a PDS/PDSE with multiple members into another PDS/PDSE by providing the complete list of members to copy to datasets._copy zoau call.

This is some performance numbers comparing current dev copy as "old copy" and proposed solution as "new copy"

Testing copy with 30 members

old copy 474.17s, 339.17s, 310.62s, 607.13s
new copy 35.75s, 55.13s, 45.11, 51.61s

Testing copy with 100 members

old copy 1275.75s (21 mins), 1400.44s(23 mins)

new copy 47.15s, 115.71s

<!--- HINT: Include "Fixes #nnn" if you are fixing an existing issue -->

##### ISSUE TYPE
<!--- Pick one below and delete the rest -->
- Feature Pull Request

Also corrected a couple of false positives in the test cases section, one was trying to copy an empty PDS into another and was a false positive in the module code. Another, was always creating a PDSE regardless of the type described in the test.

##### COMPONENT NAME
<!--- Write the short name of the module, plugin, task or feature below -->
zos_copy