@fernandofloresg 2024-01-22
<img width="1245" alt="Screenshot 2024-01-22 at 1 24 50 PM" src="https://github.com/ansible-collections/ibm_zos_core/assets/23223623/9e0c1ee5-76e3-4a9a-8a8a-e0ac840e7be6">

<img width="307" alt="Screenshot 2024-01-22 at 1 37 12 PM" src="https://github.com/ansible-collections/ibm_zos_core/assets/23223623/92370921-cfa3-4818-b019-f1a78d823651">